refactor(#592): extract BandPlanOverlay logic to band-plan-logic.ts
Move pure functions (formatFreq, hexToRgb, computePosition,
filterOverlapping, shouldSkipFetch) and RemoteSegment interface
from inline Svelte component to a standalone TypeScript module.
Tests now import from the real module instead of re-implementing.
